Add file button.py and do several important things
The important things are:
Adjust update() method of all classes to handle events.
All events are expexted to be passed to the update() method.
Add draw() method to all classes derived from Widget.
Add class Button in file button.py.
Do other things to make it work.

The target is:
In program that will use this library,
use just constructor method to create widget
and then just the update() method to make widget react on events
and then just the draw() method to draw the widget on screen surface.
Before this change user had to know and use more methods to react on events.

Also do several minor things:
Use True and False values instead of 1 and 0.