Add class Window
Code is based on Poutine but modified a lot.

Class Window serves as a container for other widget classes.
Constructor method is called with width, height and backgroundColor arguments.
After constructing it should be positioned with rect.move_ip(x,y).
And also other widgets should be added using the add(widget) method.
To update the instance with events the update(event) method should be called.
To draw the instance on screen or other pygame Surface
the draw(surface) method should be called.
Instances of classes derived from class Widget
including another Window instance
can be added to instance of Window using the add(widget) method.