Modify class Button to react on event MOUSEBUTTONUP istead of reacting on event MOUSEBUTTONDOWN.
The project Place Pictures desires such a modification in order to fix a bug.
The bug in Place Pictures was:
When user clicked with mouse on Button instance
contained within Window instance and so the Window instance was removed,
also interaction with a Picture instance was made.
In other words selected picture moved to cursor position
when user pressed OK button in a dialog.