Change import statements
Change import statements and so also change calls of the imported modules.

Do this to make this FunnyGUI work again.
The cause for that FunnyGUI stopped working is not clear and understood to me.

Describe when projects using FunnyGUI stopped working because of FunnyGUI:
I moved the project files from subfolder of the project place_pictures and this caused FunnyGUI to stop working.
To be more precise I moved the folder FunnyGUI from folder 'place_pictures' to another folder
and I made a symbolic link to FunnyGUI in folder 'place_pictures/lib'
and I expected it to continue working but it did not.
I do not remember the exact error message that Python interpreter displayed
but there was something wrong with importing.

It seems I do not understand the import system in Python 3 .
I read about the import system in Python 3 related to my situation,
but I have not found an explanation that I would understand.
I came to this correction that made FunnyGUI work again by trial and error.